Rename package_names to vnc_package_names
I ran into an issue where the wrong packages were being installed
because another role was setting the fact named package_names.  That
fact was overriding the local variable.

I will fix the other roles, but for now I can simply change the name
of the variable used in this role.