Add a Kali-specific set of installation tasks
Version 1.11.0+dfsg-1 of the Debian package tigervnc-standalone-server
has a bug that breaks functionality for the COOL. This bug has already
been fixed in the upstream TigerVNC repo, so I submitted a patch to
Debian. The patch has been accepted, and is currently available in
Debian Unstable (codenamed sid) as version 1.11.0+dfsg-2. That version
should be available in Debian Testing (codenamed bullseye, upon which
Kali Linux is based) in a few weeks.

As a temporary workaround, until the updated package is available in
Debian Testing, I am modifying this Ansible role to install version
1.10.1+dfsg-9 directly via dpkg for Kali Linux. This workaround should
be removed when version 1.11.0+dfsg-2 becomes available in Debian
Testing.

See #15 and cisagov/cool-system#183 for more details.